Feat : Integrated login into Homepage
