+++ .env
@@ -2,6 +2,6 @@ PORT=8080
 DB_URL_LOCAL=your PSQL Database URL 
 NODE_ENV=development
 # for new token visit: https://developer.spotify.com/console/get-playlist-tracks/
-TOKEN_SPOTIFY=BQAVp4UX_h3_V_-aQJxIos8NH3L87dbw08Jf55W1uHQgb9_xF_o1pOhFuHMAls1eCZu3jVvjs7QOAzY4X5G2wVjuzJeWTQLfthi2ZS_iFQubkQgShw_42dGLBQqjUNiBhG9teSd_uMeVT8HUnGr81-YENynPSNyvBqMRuiQGEHeE0x7T8PznKfyULqEzGV9a210
+TOKEN_SPOTIFY=BQAFU7eVNT5ErRy0Kk8aXj8-caaxbZB3i8FlY6zBHT65dYH460UukDB_K69Tzf7gGqweWe08rwXNjWO5YxSoy4TlHEmwNj3dHLOSeuspB_T1qfDxsOOioqAJY6dBFEZtXXxOZlpZlWPdiPq4UvBoPHr7OjVVQjZ2N-D31lsIjmCB91xW-gesrR6gJWduLXFSx3w
 CLIENT_ID = 5937b61f9fc14d478ee7b1f4328f7fe3
 CLIENT_SECRET = d92a0b5fd9ba4eb5b4889967752f81fd
